update control info in dashboard

+++ src/main/java/springboot/centralizedsystem/controllers/SubmissionController.java
@@ -33,7 +33,7 @@ public class SubmissionController extends BaseController {
         User user = SessionUtils.getUser(session);
         String token = user.getToken();
 
-        int sizeListForms = submissionService.countSubmissions(token, path);
+        long sizeListForms = submissionService.countSubmissions(token, path);
         int currPage = Integer.parseInt(page);
         int totalPages = (int) Math.ceil((float) sizeListForms / Configs.NUMBER_ROWS_PER_PAGE);
         model.addAttribute("currPage", currPage);
