+++ src/main/java/springboot/centralizedsystem/services/SubmissionService.java
@@ -6,7 +6,7 @@ public interface SubmissionService {
 
     ResponseEntity<String> findSubmissionsByPage(String token, String path, int page);
 
-    int countSubmissions(String token, String path);
+    long countSubmissions(String token, String path);
 
     ResponseEntity<String> findAllSubmissions(String token, String path, boolean isGetOnlyData);
 }
